Update script.js
Updated script to add orderCutoffTime in UTC to determine if order was placed after day's order were shipped. If order was placed too late, add an extra day to shipping.
Also added code to determine if the day is Saturday (add 2 days to shipping) or Sunday (add 1 day to shipping).